Push to private repository
ebay is an online auction and shopping website in which people buy and
sell variety of goods. The eBay prototype i.e this application is built
on node.js with angular.js, HTML5, Bootstrap and CSS as front-end and
MySQL as database. The system is designed for online shopping for users
to sell and buy products both with and without bidding. Bidding takes
place over 4 days and the person who bids highest value get possession
of the product. User can shop anytime, anywhere from our site and can
save lists, see their past purchases, and product displayed for selling.